# auth-service/src/main/java/org/example/dto/AuthResponse.java
package org.example.dto;

public record AuthResponse(
        String accessToken,
        String refreshToken,
        String tokenType,
        long expiresIn,
        UserDto user
) {

    private static final String BEARER = "Bearer";

    public static AuthResponse of(String accessToken, String refreshToken, long expiresIn, UserDto user) {
        return new AuthResponse(accessToken, refreshToken, BEARER, expiresIn, user);
    }
}