# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Финпульс — API учёта личных и семейных финансов
  version: 1.0.0
  description: |
    Контракт backend-системы учёта личных и семейных финансов.

    Архитектура: модульный монолит + выделенные сервисы **auth** и **reports**,
    за единой точкой входа (API Gateway). Все защищённые эндпоинты используют
    JWT (Bearer). Пользователь видит только свои данные и данные семей, в которых состоит.

    Ролевая модель внутри семьи: `OWNER` (админ), `MEMBER` (участник), `VIEWER` (просмотр).
  contact:
    name: Финпульс
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный API Gateway
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Пользователь создан, выданы токены
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: Email уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход по email и паролю
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access-токена по refresh-токену
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /families/{familyId}/members:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags: [Members]
      summary: Участники семьи
      operationId: listMembers
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /families/{familyId}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - name: memberId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Members]
      summary: Изменить роль участника (только OWNER)
      operationId: updateMemberRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Members]
      summary: Удалить участника (только OWNER)
      operationId: removeMember
      responses:
        '204': { description: Удалён }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /families/{familyId}/invitations:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags: [Members]
      summary: Приглашения семьи
      operationId: listInvitations
      responses:
        '200':
          description: Список приглашений
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invitation' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Members]
      summary: Пригласить участника по email (только OWNER)
      operationId: createInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '201':
          description: Приглашение создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invitation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /families/{familyId}/invitations/{invitationId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - name: invitationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Members]
      summary: Отменить приглашение (только OWNER)
      operationId: cancelInvitation
      responses:
        '204': { description: Отменено }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /invitations:
    get:
      tags: [Members]
      summary: Приглашения, адресованные текущему пользователю
      operationId: listMyInvitations
      responses:
        '200':
          description: Список входящих приглашений
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MyInvitation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /invitations/{invitationId}/accept:
    parameters:
      - name: invitationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Members]
      summary: Принять приглашение (пользователь вступает в семью)
      operationId: acceptInvitation
      responses:
        '200':
          description: Семья, в которую вступил пользователь
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Family' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /invitations/{invitationId}/decline:
    parameters:
      - name: invitationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Members]
      summary: Отклонить приглашение
      operationId: declineInvitation
      responses:
        '204': { description: Отклонено }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /operations:
    get:
      tags: [Operations]
      summary: Список операций с фильтрами и пагинацией
      operationId: listOperations
      parameters:
        - name: familyId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/OperationType' }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: memberId
          in: query
          description: Фильтр по участнику (userId)
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: { type: string, default: 'date,desc' }
      responses:
        '200':
          description: Страница операций
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Operations]
      summary: Добавить операцию (доход или расход)
      operationId: createOperation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperationRequest' }
      responses:
        '201':
          description: Создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /operations/{operationId}:
    parameters:
      - name: operationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Operations]
      summary: Получить операцию
      operationId: getOperation
      responses:
        '200':
          description: Операция
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Operations]
      summary: Изменить операцию
      operationId: updateOperation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperationRequest' }
      responses:
        '200':
          description: Обновлена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Operations]
      summary: Удалить операцию
      operationId: deleteOperation
      responses:
        '204': { description: Удалена }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Нет доступа к ресурсу
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Некорректные входные данные
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
  schemas:
    Role:
      type: string
      enum: [OWNER, MEMBER, VIEWER]
      description: Роль участника в семье
    OperationType:
      type: string
      enum: [INCOME, EXPENSE]
...
    Currency:
      type: string
      enum: [RUB, USD, EUR, KZT, GEL]
      default: RUB
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        name: { type: string, minLength: 1, maxLength: 100 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, description: Время жизни access-токена, сек }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        defaultCurrency: { $ref: '#/components/schemas/Currency' }
        createdAt: { type: string, format: date-time }
...
    Family:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        balance: { type: number, format: double, description: Текущий общий баланс }
        membersCount: { type: integer }
        myRole: { $ref: '#/components/schemas/Role' }
        createdAt: { type: string, format: date-time }
...
    Member:
      type: object
      properties:
        id: { type: string, format: uuid, description: id записи участия }
        userId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        color: { type: string, description: HEX-цвет аватара, example: '#185FA5' }
        joinedAt: { type: string, format: date-time }
    Invitation:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        status: { type: string, enum: [PENDING, ACCEPTED, EXPIRED, CANCELLED] }
        createdAt: { type: string, format: date-time }
    MyInvitation:
      type: object
      description: Приглашение со стороны получателя — для экрана «принять/отклонить»
      properties:
        id: { type: string, format: uuid }
        familyId: { type: string, format: uuid }
        familyName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        invitedBy: { type: string, description: Имя пригласившего }
        createdAt: { type: string, format: date-time }
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        familyId: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: '#/components/schemas/OperationType' }
        icon: { type: string, description: Имя иконки, example: 'shopping-cart' }
        color: { type: string, example: '#1D9E75' }
        operationsCount: { type: integer, readOnly: true }
        total: { type: number, format: double, readOnly: true }
...
    Operation:
      type: object
      properties:
        id: { type: string, format: uuid }
        familyId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/OperationType' }
        amount: { type: number, format: double, description: Положительное значение }
        currency: { $ref: '#/components/schemas/Currency' }
        date: { type: string, format: date }
        description: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        member: { $ref: '#/components/schemas/Member' }
        createdAt: { type: string, format: date-time }
    OperationRequest:
      type: object
      required: [familyId, type, amount, date, categoryId]
      properties:
        familyId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/OperationType' }
        amount:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
          description: Сумма операции, строго больше нуля
        currency: { $ref: '#/components/schemas/Currency' }
        date: { type: string, format: date }
        description: { type: string, maxLength: 255 }
        categoryId: { type: string, format: uuid }
        memberId:
          type: string
          format: uuid
          description: userId участника; по умолчанию — текущий пользователь
    OperationPage:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Operation' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        totalIncome: { type: number, format: double }
        totalExpense: { type: number, format: double }
...
    Error:
      type: object
      description: Единый формат ошибки (RFC-7807-подобный)
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        code: { type: string, example: OPERATION_NOT_FOUND }
        message: { type: string, example: Операция не найдена }
        path: { type: string }
      required: [status, code, message]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string, example: amount }
                  message: { type: string, example: должно быть больше 0 }
security:
  - bearerAuth: []